+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/facade/DBVariantContextAnnotatorFactory.java
@@ -144,4 +144,23 @@ public class DBVariantContextAnnotatorFactory {
 		return new DBVariantContextAnnotator(new CosmicAnnotationDriver(tsvPath, fastaRefPath, options), options);
 	}
 
+
+	/**
+	 * Construct Generic VCF {@link VariantContext} annotator factory.
+	 *
+	 * @param vcfDBPath
+	 *            Path to gzip-compressed, normalized and tbi-indexed VCF file to use for the annotation
+	 * @param fastaRefPath
+	 *            Path to reference FASTA file
+	 * @param options
+	 *            Configuration for the variant context annotation
+	 * @return Preconfigured {@link DBVariantContextAnnotator} object
+	 * @throws JannovarVarDBException
+	 *             on problems loading the resources
+	 */
+	public DBVariantContextAnnotator constructGenericVCF(String vcfDBPath, String fastaRefPath, DBAnnotationOptions options)
+			throws JannovarVarDBException {
+		return new DBVariantContextAnnotator(new CosmicAnnotationDriver(vcfDBPath, fastaRefPath, options), options);
+	}
+
 }